test: add persistent service test
KetchainStoreTests, UserDefaultsStoreTests
PersistentServiceTestable 프로토콜을 채택한 클래스를 추가하여 각 저장소의 CRUD 동작을 테스트합니다.
따로 처리하지 않을 unhandledError, unexpectedData를 제외한 notFound, duplicateItem 에러에 대한 테스트를 합니다.

Refs: #31
